in add note and edit note changed destination coordinates to destination name and made separate file for both.
